refactor(agentstatus): extract Cursor auth probing helpers
Move Cursor auth parsing into cursor_auth.go so service_helpers.go
stays within the repository file-length limit.
